refactor: centralize cache utilities
