+++ src/tnfr/helpers/node_cache.py
@@ -1,20 +1,18 @@
 from __future__ import annotations
 
-import hashlib
 from dataclasses import dataclass
-from functools import lru_cache
-from typing import Any, Iterable
+from typing import Any
 
 import networkx as nx  # type: ignore[import-untyped]
 
 from ..graph_utils import get_graph, get_graph_mapping
-from ..json_utils import json_dumps
-from ..logging_utils import get_logger
-
-logger = get_logger(__name__)
-
-# Key used to store the node set checksum in a graph's ``graph`` attribute.
-NODE_SET_CHECKSUM_KEY = "_node_set_checksum_cache"
+from .cache_utils import (
+    NODE_SET_CHECKSUM_KEY,
+    node_set_checksum,
+    stable_json,
+    _node_repr,
+    _node_repr_digest,
+)
 
 __all__ = (
     "NODE_SET_CHECKSUM_KEY",
@@ -26,57 +24,6 @@ __all__ = (
     "ensure_node_index_map",
     "ensure_node_offset_map",
 )
-
-
-def stable_json(obj: Any) -> str:
-    """Return a JSON string with deterministic ordering for ``obj``."""
-    return json_dumps(
-        obj,
-        sort_keys=True,
-        ensure_ascii=False,
-        to_bytes=False,
-    )
-
-
-@lru_cache(maxsize=1024)
-def _node_repr_digest(obj: Any) -> tuple[str, bytes]:
-    """Return cached stable representation and digest for ``obj``.
-
-    This single helper centralises caching for node representations and their
-    digests, ensuring both values stay in sync.
-    """
-    repr_ = stable_json(obj)
-    digest = hashlib.blake2b(repr_.encode("utf-8"), digest_size=16).digest()
-    return repr_, digest
-
-
-def clear_node_repr_cache() -> None:
-    """Clear cached node representations used for checksums."""
-    _node_repr_digest.cache_clear()
-
-
-def _node_repr(n: Any) -> str:
-    """Stable representation for node hashing and sorting."""
-    return _node_repr_digest(n)[0]
-
-def _iter_node_digests(
-    nodes: Iterable[Any], *, presorted: bool
-) -> Iterable[bytes]:
-    """Yield node digests in a deterministic order.
-
-    When ``presorted`` is ``True`` the nodes are assumed to already be sorted
-    in a stable manner and their digests are yielded directly. Otherwise,
-    the tuple of representation and digest provided by
-    :func:`_node_repr_digest` is used to avoid redundant computation.
-    """
-    if presorted:
-        for node in nodes:
-            yield _node_repr_digest(node)[1]
-    else:
-        for _, digest in sorted(
-            (_node_repr_digest(n) for n in nodes), key=lambda x: x[0]
-        ):
-            yield digest
 
 
 @dataclass(slots=True)
@@ -92,75 +39,6 @@ class NodeCache:
     @property
     def n(self) -> int:
         return len(self.nodes)
-
-
-def _node_set_checksum_no_nodes(
-    G: nx.Graph,
-    graph: Any,
-    *,
-    presorted: bool,
-    store: bool,
-) -> str:
-    """Checksum helper when no explicit node set is provided.
-
-    This isolates the flow that hashes the whole graph, preventing duplicate
-    lookups and repeated ``current_nodes`` assignments.
-    """
-    nodes_view = G.nodes()
-    current_nodes = frozenset(nodes_view)
-    cached = graph.get(NODE_SET_CHECKSUM_KEY)
-    if cached and len(cached) == 3 and cached[2] == current_nodes:
-        return cached[1]
-
-    hasher = hashlib.blake2b(digest_size=16)
-    for digest in _iter_node_digests(nodes_view, presorted=presorted):
-        hasher.update(digest)
-
-    checksum = hasher.hexdigest()
-    if store:
-        token = checksum[:16]
-        if cached and cached[0] == token:
-            return cached[1]
-        graph[NODE_SET_CHECKSUM_KEY] = (token, checksum, current_nodes)
-    return checksum
-
-
-def node_set_checksum(
-    G: nx.Graph,
-    nodes: Iterable[Any] | None = None,
-    *,
-    presorted: bool = False,
-    store: bool = True,
-) -> str:
-    """Return a BLAKE2b checksum of ``G``'s node set.
-
-    Nodes are serialised using :func:`_node_repr`. The helper
-    :func:`_iter_node_digests` yields their digests in a deterministic order,
-    handling the ``presorted`` and unsorted cases. When ``store`` is ``True``
-    the final checksum is cached under ``NODE_SET_CHECKSUM_KEY`` to avoid
-    recomputation for unchanged graphs.
-    """
-    graph = get_graph(G)
-    if nodes is None:
-        return _node_set_checksum_no_nodes(
-            G, graph, presorted=presorted, store=store
-        )
-
-    hasher = hashlib.blake2b(digest_size=16)
-
-    # Generate digests in stable order; `_iter_node_digests` sorts when needed
-    # unless `presorted` indicates the nodes are already ordered.
-    for digest in _iter_node_digests(nodes, presorted=presorted):
-        hasher.update(digest)
-
-    checksum = hasher.hexdigest()
-    if store:
-        token = checksum[:16]
-        cached = graph.get(NODE_SET_CHECKSUM_KEY)
-        if cached and cached[0] == token:
-            return cached[1]
-        graph[NODE_SET_CHECKSUM_KEY] = (token, checksum)
-    return checksum
 
 
 def _update_node_cache(
